term: first ever "app" to pull things together
This adds the initial draft of the terminal emulator. Right now this is
built around its own thing, and doesn't inherit from the heirarchy. This
is super stripped down -- basically only reads input from the pty, and
output goes out properly as well.

This will need a complete rewrite to actually work.